requirements.txt
Django==4.0.6
django-allauth==0.51.0
django-summernote==0.8.20.0
djangorestframework==3.14.0
gunicorn

# for postgres
django-environ~=0.4.5
dj-database-url==0.5.0
python-dotenv
pytest-django
psycopg2-binary
django[argon2]